switched Integer to Double for vehicle classes

+++ vehicle-consumption/src/main/java/com/ecotrekker/vehicleconsumption/messages/VehicleConsumptionReply.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import lombok.Getter;
 import lombok.Setter;
 
 @JsonInclude(Include.NON_NULL)
@@ -23,19 +22,19 @@ public class VehicleConsumptionReply {
 
     @Setter
     @JsonAlias({ "kwh", "consumption", "consum_kwh_m", "consumptionKwhM" })
-    private Integer consumptionKwhM;
+    private Double consumptionKwhM;
 
     @JsonProperty("consum_kwh_m")
-    public Integer getConsumptionKwhM() {
+    public Double getConsumptionKwhM() {
         return consumptionKwhM;
     }
 
     @Setter
     @JsonAlias({ "emissions", "co2", "co2_per_m", "co2M" })
-    private Integer co2M;
+    private Double co2M;
 
     @JsonProperty("co2_per_m")
-    public Integer getCo2M() {
+    public Double getCo2M() {
         return co2M;
     }
 
@@ -43,7 +42,7 @@ public class VehicleConsumptionReply {
         super();
     }
 
-    public VehicleConsumptionReply(String vehicle_name, Integer consumption_kwh_m, Integer co2_m) {
+    public VehicleConsumptionReply(String vehicle_name, Double consumption_kwh_m, Double co2_m) {
         this.setVehicleName(vehicle_name);
         this.setConsumptionKwhM(consumption_kwh_m);
         this.setCo2M(co2_m);
